+++ src/BloomExe/CollectionTab/LibraryListView.cs
@@ -60,6 +60,11 @@ namespace Bloom.CollectionTab
 			_sourceBooksFlow.Controls.Clear();
 			_sourceBooksFlow.HorizontalScroll.Visible = false;
 
+			if (!_model.ShowSourceCollections)
+			{
+				splitContainer1.Panel2Collapsed = true;
+			}
+
 			_headerFont = new Font(SystemFonts.DialogFont.FontFamily, (float)10.0, FontStyle.Bold);
 			_editableBookFont = new Font(SystemFonts.DialogFont.FontFamily, (float)9.0);//, FontStyle.Bold);
 			_collectionBookFont = new Font(SystemFonts.DialogFont.FontFamily, (float)9.0);
@@ -188,6 +193,23 @@ namespace Bloom.CollectionTab
 
 		private void LoadSourceCollectionButtons()
 		{
+			if (!_model.ShowSourceCollections)
+			{
+				_sourceBooksFlow.Visible = false;
+				string lockNotice = L10NSharp.LocalizationManager.GetString("CollectionTab.bookSourcesLockNotice",
+																			   "This collection is locked, so new books cannot be added/removed.");
+
+				var lockNoticeLabel = new Label()
+					{
+						Text = lockNotice,
+						Size = new Size(_primaryCollectionFlow.Width - 20, 15),
+						ForeColor = Palette.TextAgainstDarkBackground,
+						Padding = new Padding(10, 0, 0, 0)
+					};
+				_primaryCollectionFlow.Controls.Add(lockNoticeLabel);
+				return;
+			}
+
 			var collections = _model.GetBookCollections();
 			//without this guy, the FLowLayoutPanel uses the height of a button, on *the next row*, for the height of this row!
 			var invisibleHackPartner = new Label() {Text = "", Width = 0};
@@ -200,11 +222,12 @@ namespace Bloom.CollectionTab
 																				"Sources For New Shells");
 			string bookSourceHeading = L10NSharp.LocalizationManager.GetString("CollectionTab.bookSourceHeading",
 																			   "Sources For New Books");
-			bookSourcesHeader.Label.Text = _model.IsShellProject ? shellSourceHeading : bookSourceHeading;
+				bookSourcesHeader.Label.Text = _model.IsShellProject ? shellSourceHeading : bookSourceHeading;
 			invisibleHackPartner = new Label() {Text = "", Width = 0};
 			_sourceBooksFlow.Controls.Add(invisibleHackPartner);
 			_sourceBooksFlow.Controls.Add(bookSourcesHeader);
 			_sourceBooksFlow.SetFlowBreak(bookSourcesHeader, true);
+
 
 			foreach (BookCollection collection in collections.Skip(1))
 			{
@@ -422,9 +445,9 @@ namespace Bloom.CollectionTab
 				_lastClickTime = DateTime.Now;
 
 				_bookContextMenu.Enabled = true;
-				Debug.WriteLine("before selecting " + SelectedBook.Title);
+				//Debug.WriteLine("before selecting " + SelectedBook.Title);
 				_model.SelectBook(SelectedBook);
-				Debug.WriteLine("after selecting " + SelectedBook.Title);
+				//Debug.WriteLine("after selecting " + SelectedBook.Title);
 				//didn't help: _listView.Focus();//hack we were losing clicks
 				SelectedBook.ContentsChanged -= new EventHandler(OnContentsOfSelectedBookChanged); //in case we're already subscribed
 				SelectedBook.ContentsChanged += new EventHandler(OnContentsOfSelectedBookChanged);
@@ -556,12 +579,13 @@ namespace Bloom.CollectionTab
 		private void deleteMenuItem_Click(object sender, EventArgs e)
 		{
 			var button = AllBookButtons().FirstOrDefault(b => b.Tag == SelectedBook.BookInfo);
-			_model.DeleteBook(SelectedBook);
-			//ReloadCollectionButtons();
-			Debug.Assert(button != null && _primaryCollectionFlow.Controls.Contains(button));
-			if (button != null && _primaryCollectionFlow.Controls.Contains(button))
+			if (_model.DeleteBook(SelectedBook))
 			{
-				_primaryCollectionFlow.Controls.Remove(button);
+				Debug.Assert(button != null && _primaryCollectionFlow.Controls.Contains(button));
+				if (button != null && _primaryCollectionFlow.Controls.Contains(button))
+				{
+					_primaryCollectionFlow.Controls.Remove(button);
+				}
 			}
 		}